\docType{data}
\name{tnbc}
\alias{tnbc}
\title{Gene expression profiles in triple-negative breast cancer cell}
\description{
Gene expression profiles in triple-negative breast cancer cells with 215 observations and 337 variables. Gene expression values were randomly chosen from the original dataset. The outcome variable is \emph{subtype}.
}
\usage{
data(tnbc)
}
\examples{
data(tnbc)
}
\source{
Chen, X., Li, J., Gray, W. H., Lehmann, B. D., Bauer, J. A., Shyr, Y., & Pietenpol, J. A. (2012). TNBCtype: a subtyping tool for triple-negative breast cancer. \emph{Cancer informatics}, 11, CIN-S9983.
}
